Eliminate _reader_py.py, similar to earlier _schema_py.py removal
